wsign-compare

+++ src/engine/loot/LootManager.cpp
@@ -68,7 +68,7 @@ bool LootManager::canPlayAction(LootAction action) const {
     if (action.jumps.size() == 0)
         return false;
 
-    if (action.jumps.size() != 1 && step() < players.size())
+    if (action.jumps.size() != 1 && (uint)step() < players.size())
         return false;
     
     auto configuration = getConfiguration();
